deps, removed direct dependency on dateutil. utils, moved contents of dateutils into utils. utils, todt and ymdhms functions now require datetime.datetime objects and not just any old thing.
