+++ Pipfile
@@ -13,7 +13,6 @@ elifetools = "==0.32.0"
 isbnlib = "~=3.9"
 joblib = "~=1.2"
 jsonschema = "~=4.17"
-python-dateutil = "~=2.7"
 python-json-logger = "~=0.1.9"
 python-slugify = "~=1.2"
 pytz = "*"
