--- src/dateutils.py
@@ -1,39 +1,0 @@
-from datetime import datetime
-from dateutil import parser
-import json
-import pytz
-from rfc3339 import rfc3339
-
-def todt(val):
-    "turn almost any formatted datetime string into a UTC datetime object"
-    if val is None:
-        return None
-    dt = val
-    if not isinstance(dt, datetime):
-        dt = parser.parse(val, fuzzy=False)
-    dt.replace(microsecond=0) # not useful, never been useful, will never be useful.
-
-    if not dt.tzinfo:
-        # no timezone (naive), assume UTC and make it explicit
-        return pytz.utc.localize(dt)
-
-    else:
-        # ensure tz is UTC
-        if dt.tzinfo != pytz.utc:
-            return dt.astimezone(pytz.utc)
-    return dt
-
-def ymdhms(dt):
-    "returns an rfc3339 representation of a datetime object"
-    if dt:
-        dt = todt(dt) # convert to utc, etc
-        return rfc3339(dt, utc=True)
-
-def json_dumps(obj, **kwargs):
-    "drop-in for json.dumps that handles datetime objects."
-    def datetime_handler(obj):
-        if hasattr(obj, 'isoformat'):
-            return ymdhms(obj)
-        else:
-            raise TypeError('Object of type %s with value of %s is not JSON serializable' % (type(obj), repr(obj)))
-    return json.dumps(obj, default=datetime_handler, **kwargs)
